requirements.txt
httpx>=0.27
pydantic>=2.0
python-dotenv>=1.0
praw>=7.7
feedparser>=6.0
arxiv>=2.1
huggingface-hub>=0.20
anthropic>=0.40
tweepy>=4.14
scikit-learn>=1.4
fastapi>=0.115
uvicorn>=0.30
